+++ src/fastmcp/clients/__init__.py
@@ -1,12 +1,12 @@
 from .websocket import WebSocketClient
 from .sse import SSEClient
 from .stdio import StdioClient, UvxClient
-from .memory import InMemoryClient
+from .fastmcp_client import FastMCPClient
 
 __all__ = [
     "StdioClient",
     "SSEClient",
     "WebSocketClient",
     "UvxClient",
-    "InMemoryClient",
+    "FastMCPClient",
 ]
